Add initial values to statement() and context().
This change allows for a more terse version of simple objects to be
created.  In statement() it is now possible to pass a single required
expression and then add to it later (if desired).  In context() it is
now possible to add either a name and a value or a name and varargs
values.  This makes it possible to make a simple context in two
method calls.